WIP: Try to execute a query.
As it turned out - we need to populate the database with tables. For
that we need to execute liquibase change sets. The class that takes care
of that is Migrator. However it's implemented as an application context
on its own - not in a way to be used just as another Spring Bean. This
poses additional challenge for JUnit IT approach.